Implement fetching of last transaction for a family member
Added the functionality to fetch the last transaction for a family member. Made changes in the Vue components and Supabase event handler to support this. Now, for each family member, the system can show the points from the last transaction. Also, created a new form for adding points to a family member and updated relevant types and composable functions to accommodate these changes.